Add "persist.net.ipv6" setting to load IPV6.
The ipv6 module is probably not very useful right now, since dhcpcd
apparently doesn't support it, so it's just taking up memory.  Use
"setprop persist.net.ipv6 1" to enable it.

If dhcpcd starts to support ipv6, it should probably be responsible for
loading the module as needed.

(This instead of choosing based on device keeps the environment consistent
between devices, so if someone does add ipv6 support, it won't be broken
on G1.)